Add generic iterative machine test
