+++ tests/unit/classifier/Perceptron_unittest.cc
@@ -89,58 +89,3 @@ TEST(Perceptron, custom_hyperplane_initialization)
 	EXPECT_TRUE(perceptron_initialized->get_w().equals(weights));
 }
 
-TEST(Perceptron, continue_training_consistency)
-{
-	auto env = linear_test_env->getBinaryLabelData();
-	auto features = wrap(env->get_features_train());
-	auto labels = wrap(env->get_labels_train());
-	auto test_features = wrap(env->get_features_test());
-	auto test_labels = wrap(env->get_labels_test());
-
-	auto perceptron = some<CPerceptron>();
-	perceptron->set_labels(labels);
-	perceptron->train(features);
-
-	auto results = perceptron->apply(test_features);
-
-	index_t iter = 0;
-
-	auto perceptron_stop = some<CPerceptron>();
-
-	std::function<bool()> callback = [&iter]() {
-		if (iter >= 1)
-		{
-			get_global_signal()->get_subscriber()->on_next(SG_BLOCK_COMP);
-			return true;
-		}
-		iter++;
-		return false;
-	};
-	perceptron_stop->set_callback(callback);
-
-	perceptron_stop->set_labels(labels);
-	perceptron_stop->train(features);
-
-	ASSERT(!perceptron_stop->is_complete());
-
-	auto perceptron_one = some<CPerceptron>();
-	perceptron_one->set_labels(labels);
-	perceptron_one->put<int32_t>("max_iterations", 1);
-	perceptron_one->train(features);
-
-	auto results_one = perceptron_one->apply(test_features);
-	auto results_stop = perceptron_stop->apply(test_features);
-	EXPECT_TRUE(results_one->equals(results_stop));
-
-	perceptron_stop->set_callback(nullptr);
-	perceptron_stop->continue_train();
-
-	auto results_complete = perceptron_stop->apply(test_features);
-
-	EXPECT_TRUE(results_complete->equals(results));
-
-	SG_UNREF(results_one);
-	SG_UNREF(results_stop);
-	SG_UNREF(results);
-	SG_UNREF(results_complete);
-}
